*** empty log message ***

+++ RecoLocalTracker/SiStripZeroSuppression/interface/SiStripAPVRestorer.h
@@ -122,7 +122,7 @@ class SiStripAPVRestorer {
   double   CutToAvoidSignal_;	       // for iterative median implementation internal to APV restorer
   uint32_t nSaturatedStrip_;           // for BaselineAndSaturation inspect
   bool ApplyBaselineCleaner_;
-  float MeanCM_;
+  int32_t MeanCM_;
                     
 };
